+++ cli/src/cli_util.rs
@@ -28,7 +28,6 @@ use std::path::Path;
 use std::path::PathBuf;
 use std::rc::Rc;
 use std::slice;
-use std::str::FromStr;
 use std::sync::Arc;
 use std::time::SystemTime;
 
@@ -2994,56 +2993,6 @@ impl DiffSelector {
                 Ok(editor.edit([left_tree, &right_tree], matcher, format_instructions)?)
             }
         }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct RemoteBookmarkNamePattern {
-    pub bookmark: StringPattern,
-    pub remote: StringPattern,
-}
-
-impl FromStr for RemoteBookmarkNamePattern {
-    type Err = String;
-
-    fn from_str(src: &str) -> Result<Self, Self::Err> {
-        // The kind prefix applies to both bookmark and remote fragments. It's
-        // weird that unanchored patterns like substring:bookmark@remote is split
-        // into two, but I can't think of a better syntax.
-        // TODO: should we disable substring pattern? what if we added regex?
-        let (maybe_kind, pat) = src
-            .split_once(':')
-            .map_or((None, src), |(kind, pat)| (Some(kind), pat));
-        let to_pattern = |pat: &str| {
-            if let Some(kind) = maybe_kind {
-                StringPattern::from_str_kind(pat, kind).map_err(|err| err.to_string())
-            } else {
-                Ok(StringPattern::exact(pat))
-            }
-        };
-        // TODO: maybe reuse revset parser to handle bookmark/remote name containing @
-        let (bookmark, remote) = pat.rsplit_once('@').ok_or_else(|| {
-            "remote bookmark must be specified in bookmark@remote form".to_owned()
-        })?;
-        Ok(Self {
-            bookmark: to_pattern(bookmark)?,
-            remote: to_pattern(remote)?,
-        })
-    }
-}
-
-impl RemoteBookmarkNamePattern {
-    pub fn is_exact(&self) -> bool {
-        self.bookmark.is_exact() && self.remote.is_exact()
-    }
-}
-
-impl fmt::Display for RemoteBookmarkNamePattern {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // TODO: use revset::format_remote_symbol() if FromStr is migrated to
-        // the revset parser.
-        let Self { bookmark, remote } = self;
-        write!(f, "{bookmark}@{remote}")
     }
 }