+++ cli/src/commands/bookmark/untrack.rs
@@ -14,10 +14,10 @@
 
 use clap_complete::ArgValueCandidates;
 use itertools::Itertools as _;
+use jj_lib::str_util::RemoteBookmarkNamePattern;
 
 use super::find_trackable_remote_bookmarks;
 use crate::cli_util::CommandHelper;
-use crate::cli_util::RemoteBookmarkNamePattern;
 use crate::command_error::CommandError;
 use crate::complete;
 use crate::ui::Ui;